feat: Update ArXiv model and introduce end-to-end training for hierarchical classification
- Changed the default ArXiv model from sentence-transformers/all-MiniLM-L6-v2 to allenai/specter2_base.
- Added support for a maximum number of ArXiv records to load, defaulting to 50,000.
- Introduced learning rate parameter for transformer models.
- Implemented caching mechanism for feature extraction to improve efficiency.
- Added new end-to-end (E2E) training methods for global classifiers, including E2E and SOTA models.
- Updated dataset loading and management to accommodate new features and models.
- Enhanced training loops to support new model architectures and constraints.

+++ src/hmc/pipeline/global_classifier/core/train.py
@@ -55,7 +55,7 @@ def _run_training_loop(model, args, optimizer, criterion, to_eval):
             train_output = labels * output.double()
             train_output = get_constr_out(train_output, args.r_matrix)
             train_output = (1 - labels) * constr_output.double() + labels * train_output
-            loss = criterion(train_output[:, to_eval], labels[:, to_eval])
+            loss = criterion(train_output[:, to_eval].float(), labels[:, to_eval])
 
             if use_contrastive and doc_emb is not None:
                 loss = loss + lambda_c * global_contrastive_loss(doc_emb, labels)
